[#510] Improve customer account handling
- Expanded FormError logging to include request path and redirect info.
- Enhanced form error handling to differentiate based on redirect
status.
- Improved the ViewAccount logic to ensure Stripe customer and
subscription
  IDs are recorded and plan IDs are correctly updated.
- Consolidated and updated code to handle Stripe payment links, ensuring
new customer creation is consistent and existing customers are properly
  associated with checkout sessions.
- Moved `AuthenticateSession` from `logic.rb` to `account.rb` for better
  modularity and to keep related functionalities together.
