+++ lib/onetime/logic.rb
@@ -28,81 +28,6 @@ module Onetime
       end
     end
 
-    class AuthenticateSession < OT::Logic::Base
-      attr_reader :custid, :stay, :greenlighted
-      attr_reader :session_ttl
-
-      def process_params
-        @potential_custid = params[:u].to_s.downcase.strip
-        @passwd = self.class.normalize_password(params[:p])
-        #@stay = params[:stay].to_s == "true"
-        @stay = true # Keep sessions alive by default
-        @session_ttl = (stay ? 30.days : 20.minutes).to_i
-        if @potential_custid.to_s.index(':as:')
-          @colonelname, @potential_custid = *@potential_custid.downcase.split(':as:')
-        else
-          @potential_custid = @potential_custid.downcase if @potential_custid
-        end
-        if @passwd.to_s.empty?
-          @cust = nil
-        elsif @colonelname && OT::Customer.exists?(@colonelname) && OT::Customer.exists?(@potential_custid)
-          OT.info "[login-as-attempt] #{@colonelname} as #{@potential_custid} #{@sess.ipaddress}"
-          potential = OT::Customer.load @colonelname
-          @colonel = potential if potential.passphrase?(@passwd)
-          @cust = OT::Customer.load @potential_custid if @colonel.role?(:colonel)
-          sess['authenticated_by'] = @colonel.custid
-          OT.info "[login-as-success] #{@colonelname} as #{@potential_custid} #{@sess.ipaddress}"
-        elsif (potential = OT::Customer.load(@potential_custid))
-          @cust = potential if potential.passphrase?(@passwd)
-        end
-      end
-
-      def raise_concerns
-        limit_action :authenticate_session
-        if @cust.nil?
-          @cust ||= OT::Customer.anonymous
-          raise_form_error "Try again"
-        end
-      end
-
-      def process
-        if success?
-          @greenlighted = true
-
-          OT.info "[login-success] #{sess.short_identifier} #{cust.obscure_email} #{cust.role} (replacing sessid)"
-
-          # Create a completely new session, new id, new everything (incl
-          # cookie which the controllor will implicitly do above when it
-          # resends the cookie with the new session id).
-          sess.replace!
-
-          OT.info "[login-success] #{sess.short_identifier} #{cust.obscure_email} #{cust.role} (new sessid)"
-
-          sess.update_fields :custid => cust.custid, :authenticated => 'true'
-          sess.ttl = session_ttl if @stay
-          sess.save
-          cust.save
-
-          if OT.conf[:colonels].member?(cust.custid)
-            cust.role = :colonel
-          else
-            cust.role = :customer unless cust.role?(:customer)
-          end
-        else
-          raise_form_error "Try again"
-        end
-      end
-
-      def success?
-        !cust&.anonymous? && (cust.passphrase?(@passwd) || @colonel&.passphrase?(@passwd))
-      end
-
-      private
-      def form_fields
-        {:custid => custid}
-      end
-    end
-
     class DestroySession < OT::Logic::Base
       def process_params
       end
